# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ads Marketplace API",
    "version": "1.0.0",
    "description": "API documentation for the Ads Marketplace platform - connecting advertisers with channel owners on Telegram",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token"
      },
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Init-Data",
        "description": "Telegram Mini App init data"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
...
      "Channel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "telegramChatId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "PAUSED",
              "SUSPENDED",
              "REMOVED"
            ]
          },
          "isVerified": {
            "type": "boolean"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "icon": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "stats": {
            "type": "object",
            "nullable": true,
            "properties": {
              "subscribers": {
                "type": "integer"
              },
              "avgViews": {
                "type": "integer",
                "nullable": true
              },
              "engagementRate": {
                "type": "number",
                "nullable": true
              }
            }
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "priceAmount": {
                  "type": "string"
                },
                "priceCurrency": {
                  "type": "string"
                }
              }
            }
          },
          "ownerId": {
            "type": "string"
          },
          "completedDeals": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Listing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "adFormatId": {
            "type": "string"
          },
          "priceNanoTon": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Brief": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "advertiserId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "adFormatTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "customFormatDescription": {
            "type": "string",
            "nullable": true
          },
          "channelsLimit": {
            "type": "integer",
            "nullable": true
          },
          "targetCategories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minSubscribers": {
            "type": "integer",
            "nullable": true
          },
          "maxSubscribers": {
            "type": "integer",
            "nullable": true
          },
          "minAvgViews": {
            "type": "integer",
            "nullable": true
          },
          "budgetMin": {
            "type": "string",
            "nullable": true
          },
          "budgetMax": {
            "type": "string",
            "nullable": true
          },
          "totalBudget": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "desiredStartDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "desiredEndDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "flexibility": {
            "type": "string",
            "enum": [
              "STRICT",
              "FLEXIBLE",
              "ANYTIME"
            ]
          },
          "hasCreative": {
            "type": "boolean"
          },
          "creativeGuidelines": {
            "type": "string",
            "nullable": true
          },
          "applicationCount": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "COMPLETED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/briefs": {
      "get": {
        "tags": [
          "Briefs"
        ],
        "summary": "Browse active briefs",
        "description": "Browse advertiser briefs for channel owners to apply",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minBudget",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "maxBudget",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Briefs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "briefs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Brief"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Briefs"
        ],
        "summary": "Create a new brief",
        "description": "Creates a new advertising brief with requirements and budget",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "adFormatTypes",
                  "currency"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "adFormatTypes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "POST",
                        "STORY",
                        "REPOST",
                        "PINNED",
                        "OTHER"
                      ]
                    }
                  },
                  "customFormatDescription": {
                    "type": "string"
                  },
                  "channelsLimit": {
                    "type": "integer"
                  },
                  "targetCategories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "targetLanguages": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "minSubscribers": {
                    "type": "integer"
                  },
                  "maxSubscribers": {
                    "type": "integer"
                  },
                  "minAvgViews": {
                    "type": "integer"
                  },
                  "budgetMin": {
                    "type": "string"
                  },
                  "budgetMax": {
                    "type": "string"
                  },
                  "totalBudget": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Brief currency (must be one of configured supported currencies)"
                  },
                  "desiredStartDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "desiredEndDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "flexibility": {
                    "type": "string",
                    "enum": [
                      "STRICT",
                      "FLEXIBLE",
                      "ANYTIME"
                    ],
                    "default": "FLEXIBLE"
                  },
                  "hasCreative": {
                    "type": "boolean",
                    "default": false
                  },
                  "creativeGuidelines": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Brief created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "brief": {
                      "$ref": "#/components/schemas/Brief"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "List all active channels",
        "description": "Browse marketplace channels with filtering and pagination",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category slug"
          },
          {
            "in": "query",
            "name": "language",
            "schema": {
              "type": "string"
            },
            "description": "Filter by language"
          },
          {
            "in": "query",
            "name": "minSubscribers",
            "schema": {
              "type": "integer"
            },
            "description": "Minimum subscribers count"
          },
          {
            "in": "query",
            "name": "maxSubscribers",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum subscribers count"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in title, username, or description"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "subscribers",
                "price",
                "recent"
              ],
              "default": "subscribers"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "List of channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "channels": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Channel"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "Register a new channel",
        "description": "Registers a new Telegram channel to the marketplace",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "telegramChatId",
                  "title"
                ],
                "properties": {
                  "telegramChatId": {
                    "type": "number",
                    "description": "Telegram chat ID of the channel"
                  },
                  "username": {
                    "type": "string",
                    "description": "Channel username"
                  },
                  "title": {
                    "type": "string",
                    "description": "Channel title"
                  },
                  "language": {
                    "type": "string",
                    "default": "en",
                    "description": "Channel language"
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Category IDs"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Channel created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "channel": {
                      "$ref": "#/components/schemas/Channel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Channel already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deals": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "Get user's deals",
        "description": "Retrieve a paginated list of deals where the authenticated user is either the advertiser or channel owner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "advertiser",
                "publisher"
              ]
            },
            "description": "Filter by user role in deals"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by deal status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved deals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deals": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        }
      },
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "Create a new deal",
        "description": "Create a new advertising deal between advertiser and channel owner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "origin",
                  "channelId",
                  "adFormatId",
                  "agreedPrice",
                  "currency"
                ],
                "properties": {
                  "origin": {
                    "type": "string",
                    "enum": [
                      "LISTING",
                      "BRIEF",
                      "DIRECT"
                    ],
                    "description": "Origin of the deal"
                  },
                  "listingId": {
                    "type": "string",
                    "description": "Listing ID if origin is LISTING"
                  },
                  "briefId": {
                    "type": "string",
                    "description": "Brief ID if origin is BRIEF"
                  },
                  "applicationId": {
                    "type": "string",
                    "description": "Application ID if origin is BRIEF"
                  },
                  "channelId": {
                    "type": "string",
                    "description": "Channel ID for the deal"
                  },
                  "adFormatId": {
                    "type": "string",
                    "description": "Ad format ID"
                  },
                  "agreedPrice": {
                    "type": "string",
                    "description": "Agreed price for the deal"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Deal currency (must be one of configured supported currencies)"
                  },
                  "scheduledTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Scheduled time for the ad"
                  },
                  "durationHours": {
                    "type": "integer",
                    "default": 24,
                    "description": "Duration in hours"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deal": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          },
          "404": {
            "description": "Channel or ad format not found"
          }
        }
      }
    },
...
    "/api/listings": {
      "get": {
        "tags": [
          "Listings"
        ],
        "summary": "Browse active listings",
        "description": "Browse marketplace listings with filtering and pagination",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "language",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "formatType",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "listings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Listing"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Listings"
        ],
        "summary": "Create a new listing",
        "description": "Creates a new advertising listing for a channel",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "channelId",
                  "adFormatId",
                  "title"
                ],
                "properties": {
                  "channelId": {
                    "type": "string"
                  },
                  "adFormatId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "customPrice": {
                    "type": "string"
                  },
                  "customCurrency": {
                    "type": "string"
                  },
                  "availableFrom": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "availableTo": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "requirements": {
                    "type": "string"
                  },
                  "restrictions": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "listing": {
                      "$ref": "#/components/schemas/Listing"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not channel owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}